fix: update frontend
