+++ frontend/package.json
@@ -28,56 +28,55 @@
     "type-check": "tsc --pretty --noEmit"
   },
   "dependencies": {
-    "@next/bundle-analyzer": "^15.5.4",
-    "@sk-web-gui/ai": "^3.0.1",
-    "@sk-web-gui/core": "^4.1.0",
+    "@next/bundle-analyzer": "^16.1.6",
+    "@sk-web-gui/ai": "^4.0.0",
+    "@sk-web-gui/core": "^4.2.0",
     "@sk-web-gui/next": "^1.2.0",
-    "@sk-web-gui/react": "^3.0.3",
-    "axios": "^1.12.2",
-    "dayjs": "^1.11.18",
-    "i18next": "^25.5.3",
+    "@sk-web-gui/react": "^3.0.7",
+    "axios": "^1.13.5",
+    "dayjs": "^1.11.19",
+    "i18next": "^25.8.4",
     "i18next-resources-to-backend": "^1.2.1",
-    "lucide-react": "^0.545.0",
-    "next": "^15.5.7",
-    "next-i18n-router": "^5.5.5",
-    "react": "19.2.1",
-    "react-dom": "19.2.1",
-    "react-i18next": "^16.3.5",
-    "sass": "^1.93.2",
-    "sharp": "^0.34.4",
+    "lucide-react": "^0.563.0",
+    "next": "^15.5.12",
+    "next-i18n-router": "^5.5.6",
+    "react": "19.2.4",
+    "react-dom": "19.2.4",
+    "react-i18next": "^15.7.4",
+    "sass": "^1.97.3",
+    "sharp": "^0.34.5",
     "underscore.string": "^3.3.6",
-    "zustand": "^5.0.8"
+    "zustand": "^5.0.11"
   },
   "devDependencies": {
-    "@babel/plugin-transform-private-methods": "^7.27.1",
-    "@babel/preset-env": "^7.28.3",
-    "@babel/preset-react": "^7.27.1",
-    "@babel/preset-typescript": "^7.27.1",
-    "@cypress/code-coverage": "^3.14.6",
-    "@eslint/js": "^9.37.0",
+    "@babel/plugin-transform-private-methods": "^7.28.6",
+    "@babel/preset-env": "^7.29.0",
+    "@babel/preset-react": "^7.28.5",
+    "@babel/preset-typescript": "^7.28.5",
+    "@cypress/code-coverage": "^3.14.7",
+    "@eslint/js": "^10.0.1",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@tailwindcss/container-queries": "^0.1.1",
     "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^16.3.0",
+    "@testing-library/react": "^16.3.2",
     "@testing-library/user-event": "^14.6.1",
     "@types/jest": "^30.0.0",
     "@types/node": "^22.15.29",
-    "@types/react": "^19.2.7",
+    "@types/react": "^19.2.13",
     "@types/react-dom": "^19.2.3",
     "@types/underscore.string": "^0.0.41",
-    "@typescript-eslint/eslint-plugin": "^8.46.0",
-    "autoprefixer": "^10.4.21",
+    "@typescript-eslint/eslint-plugin": "^8.54.0",
+    "autoprefixer": "^10.4.24",
     "babel-jest": "^30.2.0",
     "cross-env": "^10.1.0",
-    "cypress": "^15.3.0",
-    "dotenv": "^17.2.3",
-    "dotenv-cli": "^10.0.0",
-    "envalid": "^8.1.0",
-    "eslint": "^9.37.0",
-    "eslint-config-next": "^15.5.4",
+    "cypress": "^15.10.0",
+    "dotenv": "^17.2.4",
+    "dotenv-cli": "^11.0.0",
+    "envalid": "^8.1.1",
+    "eslint": "^10.0.0",
+    "eslint-config-next": "^16.1.6",
     "eslint-config-prettier": "^10.1.8",
-    "eslint-plugin-jest": "^29.0.1",
-    "eslint-plugin-react-refresh": "^0.4.23",
+    "eslint-plugin-jest": "^29.13.0",
+    "eslint-plugin-react-refresh": "^0.5.0",
     "identity-obj-proxy": "^3.0.0",
     "istanbul-merge": "^2.0.0",
     "jest": "^30.2.0",
@@ -85,13 +84,13 @@
     "postcss": "^8.5.6",
     "postcss-import": "^16.1.1",
     "postcss-nested": "^7.0.2",
-    "postcss-nesting": "^13.0.2",
-    "prettier": "^3.6.2",
-    "swagger-typescript-api": "13.0.3",
-    "tailwindcss": "^3.4.18",
-    "ts-jest": "^29.4.4",
+    "postcss-nesting": "^14.0.0",
+    "prettier": "^3.8.1",
+    "swagger-typescript-api": "13.2.17",
+    "tailwindcss": "^3.4.19",
+    "ts-jest": "^29.4.6",
     "typescript": "^5.9.3",
-    "typescript-eslint": "^8.46.0"
+    "typescript-eslint": "^8.54.0"
   },
   "browserslist": [
     "defaults and fully supports es6-module"
