+++ frontend/postcss.config.js
@@ -1,5 +1,11 @@
 // If you want to use other PostCSS plugins, see the following:
 // https://tailwindcss.com/docs/using-with-preprocessors
 module.exports = {
-  plugins: ['postcss-import', 'tailwindcss/nesting', 'tailwindcss', 'autoprefixer'],
+  plugins: [
+    'postcss-import',
+    'tailwindcss/nesting',
+    // 'postcss-nested', // or require('postcss-nesting')
+    'tailwindcss',
+    'autoprefixer',
+  ],
 };
