+++ frontend/src/app/i18nConfig.ts
@@ -1,6 +1,7 @@
 const i18nConfig = {
   locales: ['sv'],
   defaultLocale: 'sv',
+  basePath: process.env.NEXT_PUBLIC_BASE_PATH,
 };
 
 export default i18nConfig;
