+++ frontend/tsconfig.json
@@ -13,7 +13,8 @@
     "src/**/*.ts",
     "src/**/*.tsx",
     "./*.js",
-    ".next/types/**/*.ts"
+    ".next/types/**/*.ts",
+    "next.config.js"
   ],
   "exclude": ["node_modules"]
 }
